% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrSizes.R
\name{chrSizes}
\alias{chrSizes}
\title{Chromosome sizes in hg19
Returns a named vector of chromsome sizes in hg19}
\usage{
chrSizes()
}
\description{
Chromosome sizes in hg19
Returns a named vector of chromsome sizes in hg19
}
