% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterZ.R
\name{clusterZ}
\alias{clusterZ}
\title{Cluster Samples based on Abbreviated Copy-Number Profiles}
\usage{
clusterZ(z, K = 2, plotBIC = TRUE)
}
\arguments{
\item{z}{List of components \code{bins}, \code{matrix}, \code{dat.tcn},
and \code{dat.lcn}}

\item{K}{(Range of) number of clusters to consider}
}
\value{
If \code{K} is of length more than 1, ggplot object of \code{K}
       vs. BIC. If \code{K} is a single value, data frame of cluster IDs
       for samples with posterior prob.
}
\description{
Cluster Samples based on Abbreviated Copy-Number Profiles
}
