% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneD.R
\name{oneD}
\alias{oneD}
\title{Plot Average Gain/Loss Profiles from Cohort CNA data}
\usage{
oneD(bins, mar = c(1, 2.5, 1, 2.5))
}
\arguments{
\item{mar}{Base graphics margin; argument \code{mar} to \code{par()}.}
}
\description{
Plot Average Gain/Loss Profiles from Cohort CNA data
}
