% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showHeatmap.R
\name{showHeatmap}
\alias{showHeatmap}
\title{Generate heatmap of Copy-Number Profiles}
\usage{
showHeatmap(
  z,
  clusterZ = TRUE,
  K = 2,
  reorder.hc = TRUE,
  tcn.max = 6,
  margin = c(0.05, 0.75, 0.5, 2),
  mar = c(1, 2.5, 1, 2.5)
)
}
\arguments{
\item{z}{List of components \code{bins}, \code{matrix}, \code{dat.tcn},
and \code{dat.lcn}}

\item{clusterZ}{Cluster samples using abbreviated CNA features}

\item{K}{Number of clusters}

\item{reorder.hc}{Reorder samples by hierarchical clustering after grouping by
cluster}

\item{tcn.max}{Maximum total copy number to display}

\item{mar}{Base graphics margin; argument \code{mar} to \code{par()}}
}
\description{
Generate heatmap of Copy-Number Profiles
}
